requirements.txt
azure-functions
requests
twilio